+++ BxInputController/Sources/Cells/BxInputStandartEmptyHeaderFooter.swift
@@ -1,0 +1,21 @@
+//
+//  BxInputStandartEmptyHeaderFooter.swift
+//  BxInputController
+//
+//  Created by Sergey Balalaev on 16/01/17.
+//  Copyright © 2017 Byterix. All rights reserved.
+//
+
+import UIKit
+
+class BxInputStandartEmptyHeaderFooter: BxInputStandartHeaderFooter {
+
+    @IBOutlet weak var heightConstraint: NSLayoutConstraint!
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        
+    }
+
+    
+}